break test draw contexts and events into utils
