+++ test/utils/create_mock_lifecycle_context.js
@@ -1,0 +1,7 @@
+import spy from 'sinon/lib/sinon/spy'; // avoid babel-register-related error by importing only spy
+
+export default function createMockLifecycleContext() {
+  return {
+    on: spy()
+  };
+}
